[AbstractEvenMore] Abstracted options of CreateFolderInFolder
