+++ src/FolderBehaviours/FoldInFoldOpts/OptionOther.java
@@ -1,0 +1,10 @@
+package src.FolderBehaviours.FoldInFoldOpts;
+
+public class OptionOther extends AbsOpt{
+
+    @Override
+    public boolean doOption(String folderNme, String filePath) {
+        return false;
+    }
+    
+}
